* inttypes-format: (30 commits)
  Adjust static checking of TEST_SEGSZ in test.h
  Use int32_t values in testgasnet memset test
  Add %lf to the list of non-portable printf format specifiers
  shmem: use PRI* macros for formatting 64-bit types
  psm: use PRI* macros for formatting 64-bit types
  portals4: use PRI* macros for formatting 64-bit types
  pami: use PRI* macros for formatting 64-bit types
  mxm: use PRI* macros for formatting 64-bit types
  ibv: use PRI* macros for formatting 64-bit types
  gemini: use PRI* macros for formatting 64-bit types
  template: Use PRIuPTR to output uintptr_t segsize/minheapoffset
  top-level: Use PRIuPTR and uintptr_t for formatting size_t
  other: Use PRIuPTR and uintptr_t for formatting size_t
  tests: Use PRIuPTR and uintptr_t for formatting size_t
  fix AMMPI to use 64-bit stats output
  Fix overflow in timer calibration debug output
  Remove %llu and long long from conduit-independent code
  gasnet_mmap.c: Use PRIuPTR to output uintptr_t
  Remove %llu and long long from ibv-conduit
  Remove %llu and long long from mxm-conduit
  ...

+++ gasnet_pshm.c
@@ -580,8 +580,8 @@ gasneti_pshmnet_init(void *region, size_t regionlen, gasneti_pshm_rank_t pshmnod
 
   if (regionlen < (szonce + szpernode * pshmnodes))
     gasneti_fatalerror("Internal error: not enough memory for pshmnet: \n"
-                       " given %lu effective bytes, but need %lu", 
-                       (unsigned long)regionlen, (unsigned long)(szonce + szpernode * pshmnodes));
+                       " given %"PRIuPTR" effective bytes, but need %"PRIuPTR, 
+                       (uintptr_t)regionlen, (uintptr_t)(szonce + szpernode * pshmnodes));
 
   vnet = gasneti_malloc(sizeof(gasneti_pshmnet_t));
   vnet->nodecount = pshmnodes;
