+++ other/amudp/amudp_internal.h
@@ -421,7 +421,7 @@ extern int AMUDP_Info(const char *msg, ...));
     static uint64_t _cnt = 0;                                           \
     _cnt++;                                                             \
     if_pf (!(_cnt & (_cnt-1))) {                                        \
-      AMUDP_Warn("%s (%llu occurences)",msg,(unsigned long long)_cnt);  \
+      AMUDP_Warn("%s (%" PRIu64 " occurences)",msg,_cnt);               \
     }                                                                   \
   } while (0)
 #else
@@ -443,17 +443,17 @@ GASNETT_NORETURNP(AMUDP_FatalErr)
 /* memory allocation */
 static void *_AMUDP_malloc(size_t sz, const char *curloc) {
   void *ret = malloc(sz);
-  if_pf(!ret) AMUDP_FatalErr("Failed to malloc(%lu) at %s", (unsigned long)sz, curloc);
+  if_pf(!ret) AMUDP_FatalErr("Failed to malloc(%" PRIuPTR ") at %s", (uintptr_t)sz, curloc);
   return ret;
 }
 static void *_AMUDP_calloc(size_t N, size_t S, const char *curloc) {
   void *ret = calloc(N,S);
-  if_pf(!ret) AMUDP_FatalErr("Failed to calloc(%lu,%lu) at %s", (unsigned long)N, (unsigned long)S, curloc);
+  if_pf(!ret) AMUDP_FatalErr("Failed to calloc(%" PRIuPTR ",%" PRIuPTR ") at %s", (uintptr_t)N, (uintptr_t)S, curloc);
   return ret;
 }
 static void *_AMUDP_realloc(void *ptr, size_t S, const char *curloc) {
   void *ret = realloc(ptr,S);
-  if_pf(!ret) AMUDP_FatalErr("Failed to realloc(%lu) at %s", (unsigned long)S, curloc);
+  if_pf(!ret) AMUDP_FatalErr("Failed to realloc(%" PRIuPTR ") at %s", (uintptr_t)S, curloc);
   return ret;
 }
 static void _AMUDP_free(void *ptr, const char *curloc) {
