+++ psm-conduit/gasnet_core.c
@@ -266,7 +266,7 @@ static int gasnetc_init(int *argc, char ***argv) {
         if(params.max_nargs * 2 < gasnet_AMMaxArgs()) {
             char s[255] = {0};
             snprintf(s, sizeof(s), "PSM/AM reports support for %u 32-bit arguments; GASNet requires %lu\n",
-                    params.max_nargs * 2, gasnet_AMMaxArgs());
+                    params.max_nargs * 2, (long)gasnet_AMMaxArgs());
             GASNETI_RETURN_ERRR(NOT_INIT, s);
         }
 
@@ -570,8 +570,8 @@ extern int gasnetc_attach(gasnet_handlerentry_t *table, int numentries,
         uintptr_t segsize, uintptr_t minheapoffset) {
     void *segbase = NULL;
 
-    GASNETI_TRACE_PRINTF(C,("gasnetc_attach(table (%i entries), segsize=%lu, minheapoffset=%lu)",
-                numentries, (unsigned long)segsize, (unsigned long)minheapoffset));
+    GASNETI_TRACE_PRINTF(C,("gasnetc_attach(table (%i entries), segsize=%"PRIuPTR", minheapoffset=%"PRIuPTR")",
+                numentries, segsize, minheapoffset));
 
     if (!gasneti_init_done)
         GASNETI_RETURN_ERRR(NOT_INIT, "GASNet attach called before init");
